requirements.txt
Flask
Pillow
numpy
opencv-python-headless
scipy
